notifications-service/publisher/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.poetry]
name = "publisher"
version = "0.1.0"
description = "fastapi notififcations service with rabbitmq"
authors = ["Sergey Bilibin <s.bilibin2017@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "==0.103.0"
orjson = "==3.9.7"
pydantic = "==2.4.2"
pydantic-settings = "==2.0.3"
uvicorn = "==0.23.2"
gunicorn = "==21.2.0"
python-dotenv = "1.0.0"
aio-pika = "==9.3.0"
celery = "^5.3.4"
flower = "^2.0.1"
celery-redbeat = "^2.1.1"
requests = "^2.31.0"


[tool.pylint]
max-line-length = 88
disable = ["C0114", "C0116", "R0903", "C0115"]
